Refactoring so code is more intuitive
If there was more than one term, GetLastTerm() would return the last
term with a space prepended to it.  When it is called later, this was
taken into consideration so everything worked.  However, this is
potentially confusing for someone unfamiliar with the code base.  Now,
GetLastTerm() always returns the last term without prepending a space,
and the call to it prepends a space so the space is still there when
needed.