Fixed issue where v:count wasn't recognized in some versions of Vim
On Vim 7.3 with patches 1-646 from Cygwin, :normal in a plugin appears
to change v:count.

On Vim 7.2 with patches 1-445 from Debian Squeeze, :normal in a plugin
does not appear to change v:count.

The latter behavior is the one desired here.  To ensure this happens
everywhere, the v:count is stored in another variable before any :normal
commands are called.